ci: build binaries on multi-platforms
hdiutil was removed because it uses old APIs, and fails to build. there
is no reference in the commit in which it was added to package.json.
it looks to me that hdiutil is not maintained.

update electron-builder. version 14.5.x is too old and has bugs on macos
and Windows. at least, 21.2.x is required to build the code.
see https://github.com/electron-userland/electron-builder/issues/3990
for example.

for ARCH, `amd64` and `i386` are chosen because they are widely used,
and what `uname -m` returns. i'm open for alternatives.

there are too many `if: matrix.OS == 'windows'`. just to absorb path
separator difference. when Windows Subsystem for Linux, or WSL, is
available on hosted GitHub Actions runners, they can be removed. maybe,
MSYS2 has solutions, but i did not go father because i need a working
workflow first. see available software at:
https://github.com/actions/virtual-environments/blob/master/images/win/Windows2019-Readme.md